Move parsing code to a separate file.
